feat: 🎸 Clean up terminal tab (#1876)
