+++ ui/desktop/app/components/session/terminal-tabs/index.hbs
@@ -1,0 +1,30 @@
+<Hds::Tabs as |T|>
+  <T.Tab>{{t 'titles.details'}} </T.Tab>
+  <T.Tab @icon='terminal' {{on 'click' this.openTerminal}}>
+    {{t 'resources.session.shell'}}
+  </T.Tab>
+  <T.Panel>
+    {{#if (and @model.proxy_address @model.proxy_port)}}
+      <Session::ProxyUrl
+        @proxyAddress={{@model.proxy_address}}
+        @proxyPort={{@model.proxy_port}}
+        @isSSHTarget={{@model.target.isSSH}}
+      />
+    {{/if}}
+    {{#if @model.credentials}}
+      <CredentialsPanel @credentials={{@model.credentials}} />
+    {{else}}
+      <Hds::ApplicationState data-test-no-credentials as |A|>
+        <A.Header @title={{t 'resources.session.messages.connected.title'}} />
+        <A.Body
+          @text={{t
+            'resources.session.messages.connected.description'
+            targetDisplayName=@model.target.displayName
+          }}
+        />
+      </Hds::ApplicationState>
+    {{/if}}</T.Panel>
+  <T.Panel>
+    <div id='terminal-container'></div>
+  </T.Panel>
+</Hds::Tabs>
