+++ ui/desktop/app/routes/scopes/scope/projects/sessions/session.js
@@ -5,10 +5,6 @@
 
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
-import { action } from '@ember/object';
-import { Terminal } from 'xterm';
-import { FitAddon } from 'xterm-addon-fit';
-import { v4 as uuidv4 } from 'uuid';
 
 export default class ScopesScopeProjectsSessionsSessionRoute extends Route {
   // =services
@@ -27,36 +23,5 @@ export default class ScopesScopeProjectsSessionsSessionRoute extends Route {
     return this.store.findRecord('session', session_id, {
       reload: true,
     });
-  }
-
-  // =actions
-
-  @action
-  async openTerminal(model) {
-    const xterm = new Terminal();
-    const fitAddon = new FitAddon();
-    const termContainer = document.getElementById('terminal-container');
-
-    // TODO: Do we need the fit addon?
-    xterm.loadAddon(fitAddon);
-    xterm.open(termContainer);
-    fitAddon.fit();
-
-    // Generate a UUID to have the terminal handlers be unique
-    const id = uuidv4();
-
-    // Terminal is exposed by contextBridge within the preload script.
-    window.terminal.create({ id });
-    xterm.onData((data) => window.terminal.send(data, id));
-    window.terminal.receive((event, value) => {
-      xterm.write(value);
-    }, id);
-
-    if (model.target?.isSSH) {
-      const { proxy_address, proxy_port } = model;
-
-      // Send an SSH command immediately
-      window.terminal.send(`ssh ${proxy_address} -p ${proxy_port}\r`);
-    }
   }
 }
